+++ src/core/lombok/eclipse/handlers/HandleNonNull.java
@@ -91,7 +91,7 @@ public class HandleNonNull extends EclipseAnnotationHandler<NonNull> {
 		if (isGenerated(declaration)) return;
 		
 		if (declaration.isAbstract()) {
-			annotationNode.addWarning("@NonNull is meaningless on a parameter of an abstract method.");
+			// This used to be a warning, but as @NonNull also has a documentary purpose, better to not warn about this. Since 1.16.7
 			return;
 		}